[UNTESTED] H E R E   B E   D R A G O N S
So I need a sha1 (2002 smh) of the *entire* info dictionary, not just
the pieces. I think there's a bug in be_encode() but I'm too tired to
work on it rn.

Why is SHA1 shipped with openssl? I feel like it should be something
seperate, but that's C for you.