Rewrote extract_tracker_url(). It works! (there's a leak though)
The reason we were getting a segfault before we could extract the peer
information is because the peers are not stored in a simple
sub-dictionary as I had naivly assumed, but in a _list_ of
dictionaries.

So far we've only been able to get information about one peer from a
tracker, so I don't know how well this function will scale. At least
we have peer information now!

Still gotta track down a memory leak in the library's data structure
though 🙄